ref:  DeleteDialog and EditDialog components moved into dialogs folder.
